Apply way more context to calculateForProduct
The previous function was tested in a very limited context and the few tests we did were okay by accident. We now apply a lot more product data to the simulated cart as well as validate the shipping method is actually available. The previous logic did kind of work, but ignored any rules present, so we'd get shipping methods that would otherwise have been ruled out.